% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model_stan}
\alias{logistic_model_stan}
\title{Bayesian logistic model with Stan}
\usage{
logistic_model_stan(N, t, sg, og, fg_ant, fg_sd, days, ...)
}
\arguments{
\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian logistic model with Stan
}
