% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_posterior.R
\name{sg_posterior}
\alias{sg_posterior}
\title{Posterior credible quantiles for specific gravity}
\usage{
sg_posterior(stan_fit, probs = c(0.1, 0.5, 0.9))
}
\description{
Posterior credible quantiles for specific gravity
}
